Added document for decoration visibility mainly P1
This document is intended to reduce the time needed to make changes to the code, while simultaneously improving it.  It was created after learning about and making changes to the Proposition 1 code.  Specifically when making numerous changes to when different decorations (lines, points etc.) are visible or highlighted.  It’s important to note that other content (propositions, lemmas) may be setup differently.  This document can be updated as needed to reflect new insights.