Fix time slider increasing/decreasing issues
When the delta time slider was moved to the left of its maximum eg. 0.73, then when the time slider was moved to the left from its maximum, the time shown first increased then decreased (eg. 9.43 to 10.15 then back to 9.43).  This caused similar increasing/decreasing issues for the “path step” text in the Proof tab, and showing/hiding the last red force vector in the model area.