P12/13 remove point Fi, added dragging to point P
Notes from John about the needed change to P12/13...
Point P is moved via point φ. But there's no need for point φ; we may as well move point P directly. Please remove point φ and let point P take on that interactivity.
Also, there's no need to call attention to the angle φ (the angle SP makes with line MS extended to the right), so that shaded area with label φ can be removed.