DRAFT L4 Redo - Double min rect widths, swap curve handles
Some further draft improvements for the curve manipulation have been added...
-Doubled minimum rectangle width
-Curve handles get swapped when they pass each other

Some other improvements have also been included, note they are also a work in progress...
-Transform code
-Priority for point selection

Keep in mind it's unfinished, the code is currently a bit of a messy draft, and may be a bit buggy.  Overall it's mainly intended to for test purposes.