DRAFT L4 Redo - Finalizing the code for the curve manipulation
Improvements to finalize the draft code for the curve manipulation (still a work in progress)...
-implementing the slope constraint and offset to be permanent
-set the number of curve handles to be 2
-fixing issue where bottom of figure raises when a handle is close to right side
-improving code that generates/calculates the curve
-added a test constraint to prevent the curve from becoming too vertical, from passing a sloped line
-added a test option to adjust points used to generate the curve, so that it stays more rounded when both curve handles are far apart, and one of them is close to the side

Some other improvements have also been included, note they are also a work in progress...
-transform code
-dragger code

Keep in mind it's unfinished, the code is currently a bit of a messy draft, and may be a bit buggy.  Overall it's mainly intended to for test purposes.