L4: Code cleanup
Code cleanup for L4, mainly addressed TEMP notes left in the code.  Some TEMP notes still remain, which are mainly for when this code is changed/split to apply the common changes to L2/3.  There are also a few notes for a handle dragging bug.

Overall the majority of the L4 code was finalized as the different sections were worked on.  However there were some parts that tied into other parts, which were difficult to finalize until all/the majority of the L4 changes were made.  Some notes were reminders that were no longer necessary, while others were reminders to quickly tidy and finalize a few remaining parts of the code.